test: [Node] произведен рефактор настроек тестконтейнеров
