+++ node/src/test/java/com/github/jon7even/setup/GenericRepositoryTests.java
@@ -1,5 +1,6 @@
 package com.github.jon7even.setup;
 
+import com.github.jon7even.entity.user.UserEntity;
 import com.github.jon7even.repository.UserRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -7,7 +8,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.testcontainers.context.ImportTestcontainers;
-import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -17,7 +17,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
  *
  * @author Jon7even
  * @version 2.0
- * @apiNote Использует современные аннотации, которые не требуют явного указывания секретов и настроек.
+ * @apiNote использует современные аннотации, которые не требуют явного указывания секретов и настроек.
  */
 @DataJpaTest
 @Testcontainers
@@ -25,14 +25,20 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 @ExtendWith(SpringExtension.class)
 @ImportTestcontainers(ContainerPostgreSQL.class)
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
 public class GenericRepositoryTests extends TestDataFactory {
 
     @Autowired
     protected UserRepository userRepository;
 
+    protected UserEntity userInBaseOne;
+    protected UserEntity userInBaseTwo;
+    protected UserEntity userInBaseThree;
+
     @BeforeEach
-    public void setUpEntity() {
+    public void addUser() {
         initUserEntity();
+        userInBaseOne = userRepository.save(userEntityOne);
+        userInBaseTwo = userRepository.save(userEntityTwo);
+        userInBaseThree = userRepository.save(userEntityThree);
     }
 }
